Fixed two small bugs in kml_to_sqlite.py
First bug was that the lattitude and longitude in the generated database
were swapped.  This is due to the fact that they are swapped in the KML
file format.

Second bug was a spot where the document XML object was used mistakenly
instead of the placemark XML object.  A simple bug, but it meant that
all trails were inserted into the database with the same 4 point
polyline for their trail.

Also checked in the generated SQLite database as well as the source KML
file.